Refactor mac AVPlayer package structure.
Moved AVPlayer-related classes from `sharedbuffer` to `mac.avplayer` for better organization and consistency. Removed unused imports to clean up the codebase. No functional changes introduced.
